Day 08: Resonant Collinearity
Find all antinodes of each two antennas with the same frequency. Antinodes are two times the distance from one of the points as they are from the other point. Count how many distinct antinodes there are inside the map.

Do the same, but now all points on a line through two antennas is an antinode (including the antennas themselves)