Day 13: Claw contraption
Find how you can solve the claw machine by pressing the buttons A and B to get to the prize from 0,0.
You need 3 coins to press button A, 1 to press button B. How many coins do you need for each Claw machine in which it is possible to get to the prize?

The prizes are moved +10_000_000_000_000 (1E13, 10^13) in both X and Y directions.