2025-09: Movie theater
Find the maximum area of a square with two corners that are red tiles.

For part two, the red squares with lines between each consecutive one form a surface which is filled with green squares. Find the biggest area with two red squares as corners, which is made up of only red and/or green tiles.